+++ shimming-toolbox/shimmingtoolbox/files/NiftiFile.py
@@ -4,12 +4,7 @@ import nibabel as nib
 import os
 import numpy as np
 import json
-import copy
 from functools import wraps
-from shimmingtoolbox.shim.shim_utils import extend_slice
-import math
-from shimmingtoolbox.masking.threshold import threshold
-from shimmingtoolbox.coils.coordinates import resample_from_to
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -359,207 +354,6 @@ class NiftiFile:
         model = self.get_json_info('ManufacturersModelName', required=False)
         return model.replace(" ", "_") if model is not None else None
 
-
-class NiftiFieldMap(NiftiFile):
-    """NiftiFieldMap is a subclass of NiftiFile that represents a NIfTI field map file.
-    
-    It inherits all methods and properties from NiftiFile and can be used to handle field map files specifically.
-    """
-    def __init__(self, fname_nii: str, dilation_kernel_size, path_output: str = None) -> None:
-        super().__init__(fname_nii, path_output)
-        self.extended_nii = self.extend_field_map(dilation_kernel_size)
-        self.extended_data = self.extended_nii.get_fdata()
-        self.extended_affine = self.extended_nii.affine
-        self.extended_shape = self.extended_data.shape
-        
-    def extend_field_map(self, dilation_kernel_size: int) -> None:
-        """ Extend the field map to match the dilation kernel size.
-        This method checks the dimensions of the field map and extends it if necessary.
-        Args:
-            dilation_kernel_size (int): The size of the dilation kernel to extend the field map to.
-        Raises:
-            ValueError: If the field map is not 2D or 3D.
-        Returns:
-            numpy.array : The extended NIfTI image if the field map was extended, otherwise the original NIfTI image.
-        """
-        self.extended = False
-        if self.ndim != 3:
-            if self.ndim == 2:
-                self.set_nii(nib.Nifti1Image(self.data[..., np.newaxis], self.affine,
-                                                header=self.header))
-                extended_nii = self.extend_fmap_to_kernel_size(dilation_kernel_size)
-                self.extended = True
-            else:
-                raise ValueError("Fieldmap must be 2d or 3d")
-        else:
-            for i_axis in range(3):
-                if self.shape[i_axis] < dilation_kernel_size:
-                    self.extended = True
-            if self.extended:
-                extended_nii = self.extend_fmap_to_kernel_size(dilation_kernel_size)
-                
-            if logger.level <= getattr(logging, 'DEBUG') and self.extended:
-                logger.debug(f"Field map shape: {self.shape}, "
-                             f"Extended shape: {extended_nii.shape if self.extended else self.shape}")
-                nib.save(extended_nii, os.path.join(self.path_output, f"{self.filename}_extended.nii.gz"))
-
-        return extended_nii if self.extended else self.nii
-    
-    def extend_fmap_to_kernel_size(self, dilation_kernel_size, ret_location=False):
-        """
-        Load the fmap and expand its dimensions to the kernel size
-
-        Args:
-            nii_fmap_orig (nib.Nifti1Image): 3d (dim1, dim2, dim3) or 4d (dim1, dim2, dim3, t) nii to be extended
-            dilation_kernel_size: Size of the kernel
-            path_output (str): Path to save the debug output
-            ret_location (bool): If True, return the location of the original data in the new data
-        Returns:
-            nib.Nifti1Image: Nibabel object of the loaded and extended fieldmap
-        """
-
-        fieldmap_shape = self.shape[:3]
-
-        # Extend the dimensions where the kernel is bigger than the number of voxels
-        tmp_nii = copy.deepcopy(self.nii)
-        location = np.ones(self.shape)
-        for i_axis in range(len(fieldmap_shape)):
-            # If there are less voxels than the kernel size, extend in that axis
-            if fieldmap_shape[i_axis] < dilation_kernel_size:
-                diff = float(dilation_kernel_size - fieldmap_shape[i_axis])
-                n_slices_to_extend = math.ceil(diff / 2)
-                tmp_nii, location = extend_slice(tmp_nii, n_slices=n_slices_to_extend, axis=i_axis, location=location)
-
-        nii_fmap = tmp_nii
-
-        # If DEBUG, save the extended fieldmap
-        if logger.level <= getattr(logging, 'DEBUG') and self.path_output is not None:
-            fname_new_fmap = os.path.join(self.path_output, 'tmp_extended_fmap.nii.gz')
-            nib.save(nii_fmap, fname_new_fmap)
-            logger.debug(f"Extended fmap, saved the new fieldmap here: {fname_new_fmap}")
-
-        if ret_location:
-            return nii_fmap, location.astype(bool)
-
-        return nii_fmap
-
-
-class NiftiAnatomical(NiftiFile):
-    """NiftiAnatomical is a subclass of NiftiFile that represents a NIfTI anatomical file.
-    
-    It inherits all methods and properties from NiftiFile and can be used to handle anatomical files specifically.
-    """
-    def __init__(self, fname_nii: str, path_output: str = None) -> None:
-        super().__init__(fname_nii)
-        self.check_dimensions()
-        self.temporal_average()
-        
-    def temporal_average(self):
-        if self.ndim == 3:
-            pass
-        elif self.ndim == 4:
-            logger.info("Target anatomical is 4d, taking the average and converting to 3d")
-            self.set_nii(nib.Nifti1Image(np.mean(self.data, axis=3), self.affine, header=self.header))
-        else:
-            raise ValueError("Target anatomical image must be in 3d or 4d") 
-        
-    def check_dimensions(self):
-        dim_info = self.header.get_dim_info()
-        
-        if dim_info[2] is None:
-            logger.warning("The slice encoding direction is not specified in the NIfTI header, Shimming Toolbox will "
-                        "assume it is in the third dimension.")
-        else:
-            if dim_info[2] != 2:
-                # # Reorient nifti so that the slice is the last dim
-                # anat = nii_anat.get_fdata()
-                # # TODO: find index of dim_info
-                # index_in = 0
-                # index_out = 2
-                #
-                # # Swap axis in the array
-                # anat = np.swapaxes(anat, index_in, index_out)
-                #
-                # # Affine must change
-                # affine = copy.deepcopy(nii_anat.affine)
-                # affine[:, index_in] = nii_anat.affine[:, index_out]
-                # affine[:, index_out] = nii_anat.affine[:, index_in]
-                # affine[index_out, 3] = nii_anat.affine[index_in, 3]
-                # affine[index_in, 3] = nii_anat.affine[index_out, 3]
-                #
-                # nii_reorient = nib.Nifti1Image(anat, affine, header=nii_anat.header)
-                # nib.save(nii_reorient, os.path.join(path_output, 'anat_reorient.nii.gz'))
-
-                # Slice must be the 3rd dimension of the file
-                # TODO: Reorient nifti so that the slice is the 3rd dim
-                raise RuntimeError("Slice encode direction must be the 3rd dimension of the NIfTI file.") 
-    
-    @safe_getter(default_value=False)
-    def get_fat_sat_option(self) -> bool:
-        """ Check if the NIfTI file has a Fat Saturation pulse.
-
-        Returns:
-            bool: True if Fat Saturation pulse is detected, False otherwise.
-        """
-        scan_options = self.get_json_info('ScanOptions', required=False)
-        if scan_options is not None:
-            if 'FS' in scan_options:
-                logger.debug("Fat Saturation pulse detected")
-                return True
-        else:
-            logger.debug("No ScanOptions found in the JSON metadata, assuming no Fat Saturation pulse")
-        
-        return False
-
-
-class NiftiMask(NiftiFile):
-    """NiftiMask is a subclass of NiftiFile that represents a NIfTI mask file.
-    
-    It inherits all methods and properties from NiftiFile and can be used to handle mask files specifically.
-    """
-    def __init__(self, fname_nii: str, path_output: str = None) -> None:
-        super().__init__(fname_nii, path_output, json_needed=False)
-    
-    def load_mask(self, nii_anat: NiftiAnatomical):
-        """ Load a mask and resample it on the target anatomical image.
-
-        Args:
-            nii_anat (NiftiAnatomical): The target anatomical image to resample the mask on.
-
-        Raises:
-            ValueError: If the mask is not in 3D or 4D.
-        """
-        if self.ndim == 3:
-            nii_mask_anat = self.nii
-        elif self.ndim == 4:
-            logger.debug("Mask is 4d, converting to 3d")
-            tmp_3d = np.zeros(self.shape[:3])
-            n_vol = self.shape[-1]
-            # Summing over 4th dimension making sure that the max value is 1
-            for i_vol in range(self.shape[-1]):
-                tmp_3d += (self.data[..., i_vol] / self.data[..., i_vol].max())
-            # 80% of the volumes must contain the desired pixel to be included, this avoids having dead voxels in the
-            # output mask
-            tmp_3d = threshold(tmp_3d, thr=int(n_vol * 0.8))
-            nii_mask_anat = nib.Nifti1Image(tmp_3d.astype(int), nii_mask_anat.affine,
-                                            header=nii_mask_anat.header)
-            if logger.level <= getattr(logging, 'DEBUG') and self.path_output is not None:
-                nib.save(nii_mask_anat, os.path.join(self.path_output, "fig_3d_mask.nii.gz"))
-        else:
-            raise ValueError("Mask must be in 3d or 4d")
-
-        if not np.all(nii_mask_anat.shape == nii_anat.shape) or not np.all(
-                nii_mask_anat.affine == nii_anat.affine):
-            logger.debug("Resampling mask on the target anat")
-            nii_mask_anat_soft = resample_from_to(nii_mask_anat, nii_anat.nii, order=1, mode='grid-constant')
-            tmp_mask = nii_mask_anat_soft.get_fdata()
-            # Change soft mask into binary mask
-            tmp_mask = threshold(tmp_mask, thr=0.001, scaled_thr=True)
-            nii_mask_anat = nib.Nifti1Image(tmp_mask, nii_mask_anat_soft.affine, header=nii_mask_anat_soft.header)
-            if logger.level <= getattr(logging, 'DEBUG') and self.path_output is not None:
-                nib.save(nii_mask_anat, os.path.join(self.path_output, "mask_static_resampled_on_anat.nii.gz"))
-        
-        self.set_nii(nii_mask_anat)
     
 # TODO: Implement NiftiCoilProfile class
 class NiftiCoilProfile(NiftiFile):
