+++ shimming-toolbox/shimmingtoolbox/files/NiftiMask.py
@@ -1,0 +1,58 @@
+import logging
+import nibabel as nib
+import numpy as np
+from .NiftiFile import NiftiFile
+from .NiftiAnatomical import NiftiAnatomical
+from shimmingtoolbox.masking.threshold import threshold
+from shimmingtoolbox.coils.coordinates import resample_from_to
+
+logger = logging.getLogger(__name__)
+
+class NiftiMask(NiftiFile):
+    """NiftiMask is a subclass of NiftiFile that represents a NIfTI mask file.
+    
+    It inherits all methods and properties from NiftiFile and can be used to handle mask files specifically.
+    """
+    def __init__(self, fname_nii: str, path_output: str = None) -> None:
+        super().__init__(fname_nii, path_output, json_needed=False)
+    
+    def load_mask(self, nii_anat: NiftiAnatomical):
+        """ Load a mask and resample it on the target anatomical image.
+
+        Args:
+            nii_anat (NiftiAnatomical): The target anatomical image to resample the mask on.
+
+        Raises:
+            ValueError: If the mask is not in 3D or 4D.
+        """
+        if self.ndim == 3:
+            nii_mask_anat = self.nii
+        elif self.ndim == 4:
+            logger.debug("Mask is 4d, converting to 3d")
+            tmp_3d = np.zeros(self.shape[:3])
+            n_vol = self.shape[-1]
+            # Summing over 4th dimension making sure that the max value is 1
+            for i_vol in range(self.shape[-1]):
+                tmp_3d += (self.data[..., i_vol] / self.data[..., i_vol].max())
+            # 80% of the volumes must contain the desired pixel to be included, this avoids having dead voxels in the
+            # output mask
+            tmp_3d = threshold(tmp_3d, thr=int(n_vol * 0.8))
+            nii_mask_anat = nib.Nifti1Image(tmp_3d.astype(int), nii_mask_anat.affine,
+                                            header=nii_mask_anat.header)
+            if logger.level <= getattr(logging, 'DEBUG') and self.path_output is not None:
+                nib.save(nii_mask_anat, os.path.join(self.path_output, "fig_3d_mask.nii.gz"))
+        else:
+            raise ValueError("Mask must be in 3d or 4d")
+
+        if not np.all(nii_mask_anat.shape == nii_anat.shape) or not np.all(
+                nii_mask_anat.affine == nii_anat.affine):
+            logger.debug("Resampling mask on the target anat")
+            nii_mask_anat_soft = resample_from_to(nii_mask_anat, nii_anat.nii, order=1, mode='grid-constant')
+            tmp_mask = nii_mask_anat_soft.get_fdata()
+            # Change soft mask into binary mask
+            tmp_mask = threshold(tmp_mask, thr=0.001, scaled_thr=True)
+            nii_mask_anat = nib.Nifti1Image(tmp_mask, nii_mask_anat_soft.affine, header=nii_mask_anat_soft.header)
+            if logger.level <= getattr(logging, 'DEBUG') and self.path_output is not None:
+                nib.save(nii_mask_anat, os.path.join(self.path_output, "mask_static_resampled_on_anat.nii.gz"))
+        
+        self.set_nii(nii_mask_anat)
